+++ src/drivers/LinkedIn.php
@@ -1,7 +1,8 @@
 <?php
+
 /**
  * @link https://github.com/yiimaker/yii2-social-share
- * @copyright Copyright (c) 2017-2019 Yii Maker
+ * @copyright Copyright (c) 2017-2021 Volodymyr Kupriienko
  * @license BSD 3-Clause License
  */
 
@@ -12,11 +13,12 @@ use ymaker\social\share\base\AbstractDriver;
 /**
  * Driver for LinkedIn.
  *
- * @link https://linkedin.com
+ * @see https://linkedin.com
  *
  * @property bool|string $siteName
  *
  * @author Vladimir Kuprienko <vldmr.kuprienko@gmail.com>
+ *
  * @since 1.0
  */
 class LinkedIn extends AbstractDriver
@@ -25,7 +27,6 @@ class LinkedIn extends AbstractDriver
      * @var bool|string
      */
     public $siteName = false;
-
 
     /**
      * {@inheritdoc}