+++ src/drivers/Telegram.php
@@ -1,7 +1,8 @@
 <?php
+
 /**
  * @link https://github.com/yiimaker/yii2-social-share
- * @copyright Copyright (c) 2017-2019 Yii Maker
+ * @copyright Copyright (c) 2017-2021 Volodymyr Kupriienko
  * @license BSD 3-Clause License
  */
 
@@ -12,9 +13,10 @@ use ymaker\social\share\base\AbstractDriver;
 /**
  * Driver for Telegram messenger.
  *
- * @link https://telegram.org
+ * @see https://telegram.org
  *
  * @author Vladimir Kuprienko <vldmr.kuprienko@gmail.com>
+ *
  * @since 1.0
  */
 class Telegram extends AbstractDriver
@@ -23,7 +25,6 @@ class Telegram extends AbstractDriver
      * @var bool|string
      */
     public $message = false;
-
 
     /**
      * {@inheritdoc}