+++ src/drivers/WhatsApp.php
@@ -1,7 +1,8 @@
 <?php
+
 /**
  * @link https://github.com/yiimaker/yii2-social-share
- * @copyright Copyright (c) 2017-2019 Yii Maker
+ * @copyright Copyright (c) 2017-2021 Volodymyr Kupriienko
  * @license BSD 3-Clause License
  */
 
@@ -12,12 +13,13 @@ use ymaker\social\share\base\AbstractDriver;
 /**
  * Driver for WhatsApp messenger.
  *
- * @link https://www.whatsapp.com
+ * @see https://www.whatsapp.com
  *
  * WARNING: This driver works only in mobile devices
  * with installed WhatsApp client.
  *
  * @author Vladimir Kuprienko <vldmr.kuprienko@gmail.com>
+ *
  * @since 1.0
  */
 class WhatsApp extends AbstractDriver