+++ tests/_config/app.php
@@ -1,5 +1,11 @@
 <?php
 
+/**
+ * @link https://github.com/yiimaker/yii2-social-share
+ * @copyright Copyright (c) 2017-2021 Volodymyr Kupriienko
+ * @license BSD 3-Clause License
+ */
+
 return [
     'id' => 'test-app',
     'class' => yii\console\Application::class,