+++ tests/_support/Helper/Unit.php
@@ -1,11 +1,12 @@
 <?php
+
 /**
  * @link https://github.com/yiimaker/yii2-social-share
- * @copyright Copyright (c) 2017-2019 Yii Maker
+ * @copyright Copyright (c) 2017-2021 Volodymyr Kupriienko
  * @license BSD 3-Clause License
  */
 
-namespace Helper;
+namespace ymaker\social\share\tests\Helper;
 
 /**
  * Helper for unit tests.