+++ tests/drivers/TwitterTest.php
@@ -1,11 +1,12 @@
 <?php
+
 /**
  * @link https://github.com/yiimaker/yii2-social-share
- * @copyright Copyright (c) 2017-2019 Yii Maker
+ * @copyright Copyright (c) 2017-2021 Volodymyr Kupriienko
  * @license BSD 3-Clause License
  */
 
-namespace ymaker\social\share\tests\unit\drivers;
+namespace ymaker\social\share\tests\drivers;
 
 use Codeception\Test\Unit;
 use Yii;
@@ -15,9 +16,10 @@ use ymaker\social\share\drivers\Twitter;
 /**
  * Test case for [[Twitter]] driver.
  *
- * @property-read \UnitTester $tester
+ * @property \ymaker\social\share\tests\UnitTester $tester
  *
  * @author Vladimir Kuprienko <vldmr.kuprienko@gmail.com>
+ *
  * @since 1.4.0
  */
 class TwitterTest extends Unit
@@ -35,7 +37,7 @@ class TwitterTest extends Unit
             . '&text=' . AbstractDriver::encodeData($description)
             . '&via=' . AbstractDriver::encodeData($account);
 
-        self::assertEquals($expected, $driver->getLink());
+        static::assertEquals($expected, $driver->getLink());
     }
 
     public function testRegisteredMetaTags()
@@ -58,6 +60,6 @@ class TwitterTest extends Unit
             $this->tester->metaTag('twitter:image', $imageUrl),
         ];
 
-        self::assertEquals($expected, \array_values(Yii::$app->getView()->metaTags));
+        static::assertEquals($expected, \array_values(Yii::$app->getView()->metaTags));
     }
 }