+++ tests/widgets/SocialShareTest.php
@@ -1,11 +1,12 @@
 <?php
+
 /**
  * @link https://github.com/yiimaker/yii2-social-share
- * @copyright Copyright (c) 2017-2019 Yii Maker
+ * @copyright Copyright (c) 2017-2021 Volodymyr Kupriienko
  * @license BSD 3-Clause License
  */
 
-namespace ymaker\social\share\tests\unit\widgets;
+namespace ymaker\social\share\tests\widgets;
 
 use Codeception\Test\Unit;
 use Yii;
@@ -15,9 +16,10 @@ use ymaker\social\share\widgets\SocialShare;
 /**
  * Test case for  [[ymaker\social\share\widgets\SocialShare]].
  *
- * @property-read \UnitTester $tester
+ * @property \ymaker\social\share\tests\UnitTester $tester
  *
  * @author Vladimir Kuprienko <vldmr.kuprienko@gmail.com>
+ *
  * @since 1.0
  */
 class SocialShareTest extends Unit
@@ -25,7 +27,6 @@ class SocialShareTest extends Unit
     const DEFAULT_CONFIG_CONFIGURATOR_ID = 'defaultConfig';
     const DEFAULT_ICONS_CONFIGURATOR_ID = 'defaultIcons';
 
-    private $vk = 'http://vk.com/share.php?url=test+url&amp;title=test+title&amp;description=test+description&amp;image=test+image+url';
     private $facebook = 'http://www.facebook.com/sharer.php?u=test+url';
     private $twitter = 'http://twitter.com/share?url=test+url&amp;text=test+description';
     private $googlePlus = 'https://plusone.google.com/_/+1/confirm?hl=en&amp;url=test+url';
@@ -37,10 +38,8 @@ class SocialShareTest extends Unit
     private $gmail = 'https://mail.google.com/mail/?view=cm&amp;fs=1&amp;su=test+title&amp;body=test+description+-+test+url';
     private $tumblr = 'https://www.tumblr.com/widgets/share/tool?canonicalUrl=test+url&amp;posttype=link&amp;caption=test+title&amp;content=test+url';
     private $yahoo = 'https://compose.mail.yahoo.com/?subject=test+title&amp;body=test+description+-+test+url';
-    private $odnoklassniki = 'https://connect.ok.ru/offer?url=test+url&amp;title=test+title&amp;description=test+description&amp;imageUrl=test+image+url';
     private $trello = 'https://trello.com/add-card?url=test+url&amp;name=test+title';
 
-    
     public function testInvalidConfigException()
     {
         $this->expectException(InvalidConfigException::class);
@@ -53,7 +52,6 @@ class SocialShareTest extends Unit
         $expectedHTML =
                     '<!--noindex-->'
                     . '<ul class="social-share">'
-                        . "<li><a href=\"$this->vk\" rel=\"noopener\" target=\"_blank\">Vkontakte</a></li>"
                         . "<li><a href=\"$this->facebook\" rel=\"noopener\" target=\"_blank\">Facebook</a></li>"
                         . "<li><a href=\"$this->twitter\" rel=\"noopener\" target=\"_blank\">Twitter</a></li>"
                         . "<li><a href=\"$this->googlePlus\" rel=\"noopener\" target=\"_blank\">Google Plus</a></li>"
@@ -65,13 +63,12 @@ class SocialShareTest extends Unit
                         . "<li><a href=\"$this->gmail\" rel=\"noopener\" target=\"_blank\">Gmail</a></li>"
                         . "<li><a href=\"$this->tumblr\" rel=\"noopener\" target=\"_blank\">Tumblr</a></li>"
                         . "<li><a href=\"$this->yahoo\" rel=\"noopener\" target=\"_blank\">Yahoo</a></li>"
-                        . "<li><a href=\"$this->odnoklassniki\" rel=\"noopener\" target=\"_blank\">Odnoklassniki</a></li>"
                         . "<li><a href=\"$this->trello\" rel=\"noopener\" target=\"_blank\">Trello</a></li>"
                     . '</ul>'
                     . '<!--/noindex-->';
         $actualHTML = $this->getActualHTML(self::DEFAULT_CONFIG_CONFIGURATOR_ID);
 
-        self::assertEquals($expectedHTML, $actualHTML, 'Widget should render share links');
+        static::assertEquals($expectedHTML, $actualHTML, 'Widget should render share links');
 
         $expectedMetaTags = [
             $this->tester->openGraphMetaTag('og:url', 'test url'),
@@ -85,7 +82,7 @@ class SocialShareTest extends Unit
             $this->tester->metaTag('twitter:image', 'test image url'),
         ];
 
-        self::assertEquals($expectedMetaTags, \array_values(Yii::$app->getView()->metaTags));
+        static::assertEquals($expectedMetaTags, \array_values(Yii::$app->getView()->metaTags));
     }
 
     public function testDefaultIcons()
@@ -93,7 +90,6 @@ class SocialShareTest extends Unit
         $expectedHTML =
             '<!--noindex-->'
             . '<ul class="social-share">'
-                . "<li><a href=\"$this->vk\" rel=\"noopener\" target=\"_blank\"><i class=\"si si-vk\"></i></a></li>"
                 . "<li><a href=\"$this->facebook\" rel=\"noopener\" target=\"_blank\"><i class=\"si si-facebook\"></i></a></li>"
                 . "<li><a href=\"$this->twitter\" rel=\"noopener\" target=\"_blank\"><i class=\"si si-twitter\"></i></a></li>"
                 . "<li><a href=\"$this->googlePlus\" rel=\"noopener\" target=\"_blank\"><i class=\"si si-google-plus\"></i></a></li>"
@@ -105,14 +101,13 @@ class SocialShareTest extends Unit
                 . "<li><a href=\"$this->gmail\" rel=\"noopener\" target=\"_blank\"><i class=\"si si-gmail\"></i></a></li>"
                 . "<li><a href=\"$this->tumblr\" rel=\"noopener\" target=\"_blank\"><i class=\"si si-tumblr\"></i></a></li>"
                 . "<li><a href=\"$this->yahoo\" rel=\"noopener\" target=\"_blank\"><i class=\"si si-yahoo\"></i></a></li>"
-                . "<li><a href=\"$this->odnoklassniki\" rel=\"noopener\" target=\"_blank\"><i class=\"si si-odnoklassniki\"></i></a></li>"
                 . "<li><a href=\"$this->trello\" rel=\"noopener\" target=\"_blank\"><i class=\"si si-trello\"></i></a></li>"
             . '</ul>'
             . '<!--/noindex-->';
 
         $actualHTML = $this->getActualHTML(self::DEFAULT_ICONS_CONFIGURATOR_ID);
 
-        self::assertEquals($expectedHTML, $actualHTML, 'Widget should render share links with default icons');
+        static::assertEquals($expectedHTML, $actualHTML, 'Widget should render share links with default icons');
     }
 
     public function testDisableMetaTags()
@@ -122,7 +117,7 @@ class SocialShareTest extends Unit
             'url' => 'test',
         ]);
 
-        self::assertSame([], Yii::$app->getView()->metaTags);
+        static::assertSame([], Yii::$app->getView()->metaTags);
     }
 
     /**
@@ -135,11 +130,11 @@ class SocialShareTest extends Unit
     private function getActualHTML($configurator)
     {
         $widget = new SocialShare([
-            'configurator'  => $configurator,
-            'url'           => 'test url',
-            'title'         => 'test title',
-            'description'   => 'test description',
-            'imageUrl'      => 'test image url',
+            'configurator' => $configurator,
+            'url' => 'test url',
+            'title' => 'test title',
+            'description' => 'test description',
+            'imageUrl' => 'test image url',
         ]);
         \ob_start();
         $widget->run();